% Auto-generated: do not edit by hand
\name{tvlwc}

\alias{tvlwc}

\title{Tvlwc component}

\description{
Tradingview Lightweight Chart object
}

\usage{
tvlwc(id=NULL, colors=NULL, data=NULL)
}

\arguments{
\item{id}{Character. The ID of this component}

\item{colors}{Named list. An object containing colors properties}

\item{data}{Unnamed list. The data for the series}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

